Setting Manage Municipalities use case

+++ src/main/java/client/model/municipalities/ManageMunicipalities.java
@@ -6,10 +6,8 @@ import java.util.ArrayList;
 
 public interface ManageMunicipalities
 {
-
   public void addMunicipality(Municipality municipality);
   public void addRegionalAdmin(RegionalAdmin regionalAdmin, Municipality municipality);
   public ArrayList<Municipality> getMunicipalities();
-  public Municipality getMunicipality(int id);
-
+  public Municipality getMunicipality(Long id);
 }
